Extract prefix and name into new object
I've committed this immediately after updating the "weapon has
enchantment" test to pass the new Spell object into the enchant()
method.

I felt that "prefix" and "attribute" weren't great names; they didn't
express any kind of intent as to what they were for, or what they
related to in the "real world" (I know this is "magic", so not real, but
I'd like these names to relate to the world of the problem we're working
on).

Next step, after getting this test to pass, is to refactor.