Refactoring: Use Null object to represent no spell
The "null object" pattern is a lovely one. It allows us to represent the
case where something isn't set with an object that has exactly the same
"shape" as the thing that would be there if the property was set.

In this case, we're saying "when there isn't a spell on the weapon, the
weapon can interact with a spell-shaped object that will behave like
there isn't a spell there at all".

And that simplifies the code, removing the need to check if the spell is
None.